pyproject.toml
[tool.poetry]
name = "xhup-club-api"
version = "0.0.7"
description = ""
authors = ["ryan4yin <xiaoyin_c@qq.com>"]
license = "MIT"
repository = "https://github.com/coding-and-typing/xhup-club-api"

[[tool.poetry.source]]
name = "aliyun-pypi"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22.0"
lxml = "^4.6.3"
captcha = "^0.3"
beautifulsoup4 = "^4.8.2"
flask = "^1.1.1"
flask-smorest = "^0.18.4"
flask-limiter = "^1.1.0"
flask-sockets = "^0.2.1"
flask-cors = "^3.0.8"
flask-mail = "^0.9.1"
flask-login = "^0.4.1"
flask-and-redis = "^1.0.0"
flask-rq2 = "^18.3"
flask-sqlalchemy = "^2.4.1"
flask-migrate = "^2.5.2"
pyjwt = "^1.7.1"
pymysql = "^0.9.3"
redis = "^3.3.11"
elasticsearch = "^7.1.0"
pyyaml = "^5.4.1"
pydantic = {extras = ["email"], version = "^1.8.2"}

[tool.poetry.dev-dependencies]
python-dotenv = "^0.10.3"
pytest = "^5.3.2"
pytest-cov = "^2.8.1"
codecov = "^2.0.15"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
